refactor(pwa): extract BottomSheet primitive from 3 components
